requirements.txt
Flask
gevent
gunicorn
keras
numpy
Image
h5py
pillow
tensorflow
werkzeug